+++ adminer/drivers/mysql.inc.php
@@ -1145,6 +1145,7 @@ WHERE ROUTINE_SCHEMA = DATABASE() AND ROUTINE_TYPE = '$type' AND ROUTINE_NAME = 
 	* - scheme - schemas inside a database
 	* - sequence - sequences (PostgreSQL)
 	* - single_db - there is only one database so the user is taken to it and the database select is not printed (Elasticsearch)
+	* - single_table - there is only one table in each database so the user is taken to it (Redis)
 	* - sql - the SQL command page and running multiple queries
 	* - status - the Status page
 	* - table - tables have a fixed structure, other drivers just select all columns